fix():ticket repo.js  created
