+++ backend/src/repository/UserRepository.js
@@ -9,13 +9,6 @@ export class UserRepository {
       [user.name, user.email, user.password]
     );
   }
-  async saveTicket(ticket) {
-    await db.query(
-      `INSERT INTO foxticket.TicketTypes (name, price, description, icon)
-         VALUES (?, ?, ?, ?)`,
-      [ticket.name, ticket.price, ticket.description, ticket.icon]
-    );
-  }
   
   async getUserData(username, userEmail) {
     return await db.query(
