+++ backend/src/seed.js
@@ -1,11 +1,19 @@
 import { UserRepository } from './repository/UserRepository';
+import { TicketRepository } from './repository/TicketTypesRepository';
 
-let repo = new UserRepository()
+let userRepo = new UserRepository()
+let ticketRepo = new TicketRepository()
 
 let userDummyData = {
   name: 'Lehel',
   email: 'lehel@gmail.com',
   password: 'password123',
+};
+let adminDummyData = {
+  name: 'Lehel',
+  email: 'lehel@gmail.com',
+  password: 'password123',
+  isAdmin: true
 };
 let TicketDummyData = {
   name: 'Mav',
@@ -14,6 +22,6 @@ let TicketDummyData = {
   icon: 'FaBeer'
 };
 
-
-repo.save(userDummyData);
-repo.saveTicket(TicketDummyData);
+userRepo.save(userDummyData);
+userRepo.save(adminDummyData);
+ticketRepo.saveTicket(TicketDummyData);
